move 1d processing to its own module
